Create multiple actors
This code is illustratively broken. Normally, it is bad practice to
commit broken code, however, since this is a class project in
concurrency, I have included this code to and screenshot to point out
what happens in Go when there is a data race. Specifically, when two
goroutines try to write simultaneously to the same Map.