Rework Login to use WebSocket as Client
The motivation for this large change was to allow for the Server to send
new buzzes to the client. This bidirectional communication is different
than the traditional Server-Client model of HTTP, in which a client
posts a message and wait for an OK or error message. Furthermore,
clients frequently have to poll the server for updates.

Anyway, I also found it necessary to use a sync.Mutex to prevent
concurrent writers to a socket.