Support mentions on the frontend
A mention will automatically turn into a link which, when clicked, will
display the mentioned user's buzzes. In the backend, the server is
streaming messages one by one instead of all at once as there may be
enough messages that sending in bulk would be too large or take too long
time to retrieve.