Add ability to follow
This required a change to Login in which a copy of the User data is
returned upon authentication. The WebSocket layer sends an OK response,
but then also goes through the list of followees and sends a "follow
USERNAME" for each one.

Also, I confirmed the semantics of copying a struct containing a slice
in the unit test TestCopySlice. This shows how slices are effectively
immutable as appending returns a new slice. Each element is free to be
manipulated, however, so the code still at risk of having a data-race.
The reason there isn't a data race, however, is that the Username and
password fields for a User object are never rewritten. Furthermore,
the original slice for the list of followees is not modified. So, as it
is now, the code is free of data-races, but only because of programmer
guarantees. I've placed a warning next to the risky bits in Login().