Add unfollow
I rewrote much of the accept() function to better handle the situation
where the same user is logged into multiple devices. This also
illuminated a data-race involving the client.username variable. I hadn't
noticed it before because I didn't try to switch accounts, a scenario
where I was reading and writing to client.username from two different
goroutines.

I could have used an atomic.Value, but as the documentation says, it's
preferred to use a channel. After thinking about how to get the same
behavior, I came up with the idea to turn the variable into a
string-semaphore in which goroutines need to acquire and release it.

To acquire, receive on the channel `username <- client.username`.
To release, send on the channel `client.username <- username`.

After searching the Internet, however, I found out I'm not the first to
come up with this idea.