Add support for looking up user
By changing the order of some argument checks, you can now see if a user
exists by sending the message "login USERNAME " *note the extra space at
the end.* This will return "Invalid credentials" when the user exists
instead of "Unknown user" when the user does not. This will be useful
when a client wants to search for a user by their handle without having
to add another message type.